refactor: getFells() in FellController
The way a collection of fells ire returned by the controller is much more structured now. First the service pulls the fells out of the repo, those fells are then mapped into a SimpleFellsDTO, the appropriate nav links (prev, self and next) are created and mapped to a LinksDTO, then finally a PagedCollectionDTO is assembled with the fells, links and other page meta data.