refactor: implement retrieving partial projection from Fell repo
Currently the requirement when fetching all fells from the fells repo is that we only need the fell name and os map ref (for creating the self href) and not the whole set of data associated with the entity. So now the fell controller can ask the fell service to get all fells with an offset and limit as well as a projection to control which attributes are returned.