fix: order of fells returned in response
The repository was correctly returning the fells in the required order but the controller, when transforming the results into a LinkedResultPage, was using an unordered Set which was mixing up the order of the fells. Once switching to an explicit LinkedHashSet (to preserve insertion order and uniqueness) the tests passed again.